requirements.txt
# requirements.txt
streamlit
pandas
openpyxl
google-generativeai
python-dotenv
yt-dlp 